% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfree_index_multi.R
\name{dfree_index_multi}
\alias{dfree_index_multi}
\title{Index for the MDFFSS scan procedure}
\usage{
dfree_index_multi(data, matrix_clusters)
}
\arguments{
\item{data}{List. List of the data, each element of the list corresponds to a site (or an individual), for each element each row corresponds to a variable and each column represents an observation time.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
