% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fmulti_data.R
\docType{data}
\name{fmulti_data}
\alias{fmulti_data}
\title{Multivariate functional data}
\format{
A list of 169 elements. Each element corresponds to a canton and is a matrix of 56 columns (for the 56 days of observation) and 4 rows (4 variables, in the order NO2, O3, PM10 and PM2.5).
}
\usage{
data("fmulti_data")
}
\description{
Concentrations over the time of NO2, O3, PM10 and PM2.5 from 2020/05/01 to 2020/06/25 in each canton (administrative subdivision) of Nord-Pas-de-Calais (a region from France).
}
\references{
Data from the National Air Quality Forecasting Platform www.prevair.org
}
\keyword{datasets}
