% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map_sites.R
\docType{data}
\name{map_sites}
\alias{map_sites}
\title{Spatial object corresponding to the sites of the data of the package HDSpatialScan}
\format{
A SpatialPolygonsDataFrame.
}
\usage{
data("map_sites")
}
\description{
Spatial object corresponding to the sites (169 cantons) of the data of the package HDSpatialScan.
}
\keyword{datasets}
