% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_fonc.R
\name{multi_fWMW}
\alias{multi_fWMW}
\title{Index for the NPFSS scan procedure (multivariate functional case)}
\usage{
multi_fWMW(signs, matrix_clusters)
}
\arguments{
\item{signs}{list of numeric matrices. List of nb_sites (or nb_individuals) sign matrices, the rows correspond to the variables and each column represents an observation time.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
