% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_multi.R
\name{multi_gaussian}
\alias{multi_gaussian}
\title{Index for the MG scan procedure}
\usage{
multi_gaussian(data, matrix_clusters)
}
\arguments{
\item{data}{numeric matrix. Matrix of the data, the rows correspond to the sites (or individuals) and each column represents a variable.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to minimize on the set of potential clusters, for each potential cluster
}
