% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_fonc.R
\name{multi_signs_matrix}
\alias{multi_signs_matrix}
\title{List of matrix of signs (multivariate functional data)}
\usage{
multi_signs_matrix(data)
}
\arguments{
\item{data}{list of numeric matrices. List of nb_sites (or nb_individuals) matrices of the data, the rows correspond to the variables and each column represents an observation time.}
}
\value{
list of numeric matrices.
}
\description{
This function returns the list of matrix of signs for the multivariate functional data
}
