% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_multi.R
\name{multi_WMW}
\alias{multi_WMW}
\title{Index for the MNP scan procedure}
\usage{
multi_WMW(rank_data, matrix_clusters)
}
\arguments{
\item{rank_data}{numeric matrix. Matrix of the ranks of the initial data, the rows correspond to the sites (or the individuals) and each column represents a variable.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
