% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlap.R
\name{non_overlap}
\alias{non_overlap}
\title{Return only the detected clusters with no overlapping in their order of detection}
\usage{
non_overlap(index_clusters, matrix_clusters)
}
\arguments{
\item{index_clusters}{numeric vector. The indices of the detected clusters.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. A value of 1 indicate that the site (or the individual) is in the cluster, 0 otherwise.}
}
\value{
The detecting clusters with no overlapping, in their order of detection.
}
\description{
This function allows to return only the detected clusters with no overlapping in their order of detection.
}
