% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate}
\alias{permutate}
\title{Permutates the data}
\usage{
permutate(to_permute, nb_permu)
}
\arguments{
\item{to_permute}{vector. Vector of indices we want to permute.}

\item{nb_permu}{numeric. Number of permutations.}
}
\value{
matrix. Matrix of nb_permu rows and length(to_permute) columns.
}
\description{
This function will permit to permute the data for the MC simulations
}
