% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointwise_dfree.R
\name{pointwise_dfree}
\alias{pointwise_dfree}
\title{Index for the DFFSS scan procedure}
\usage{
pointwise_dfree(data, matrix_clusters)
}
\arguments{
\item{data}{numeric matrix. Matrix of the data. The rows correspond to the sites (or the individuals) and each column represents an observation time.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
