% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointwise_wmw_multi.R
\name{pointwise_wmw_multi}
\alias{pointwise_wmw_multi}
\title{Index for the MRBFSS scan procedure}
\usage{
pointwise_wmw_multi(transform_data, matrix_clusters)
}
\arguments{
\item{transform_data}{List. List of the data transformed with the function transform_data, each element of the list corresponds to an observation time. Each row of each element is a site (or an individual), and each column represents a variable.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
