% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointwise_wmw_uni.R
\name{pointwise_wmw_uni}
\alias{pointwise_wmw_uni}
\title{Index for the URBFSS scan procedure}
\usage{
pointwise_wmw_uni(rank_data, matrix_clusters)
}
\arguments{
\item{rank_data}{matrix. Matrix of the ranks of the data for each time. Each column corresponds to an observation time and each row corresponds to a site or an individual.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric vector.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster
}
