% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriori_filtering.R
\name{post_filt_area}
\alias{post_filt_area}
\title{A posteriori filtering on the area}
\usage{
post_filt_area(mini_post, maxi_post, areas_clusters, index_clusters_temp)
}
\arguments{
\item{mini_post}{numeric. A minimum to filter the significant clusters a posteriori. The default NULL is for no filtering with a a posteriori minimum.}

\item{maxi_post}{numeric. A maximum to filter the significant clusters a posteriori. The default NULL is for no filtering with a a posteriori maximum.}

\item{areas_clusters}{numeric vector. The areas of the clusters.}

\item{index_clusters_temp}{numeric vector. The indices of the detected clusters.}
}
\value{
The detecting clusters with the a posteriori filtering.
}
\description{
This function allows the a posteriori filtering on the area.
}
