% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriori_filtering.R
\name{post_filt_nb_sites}
\alias{post_filt_nb_sites}
\title{A posteriori filtering on the number of sites/individuals}
\usage{
post_filt_nb_sites(
  mini_post,
  maxi_post,
  nb_sites,
  index_clusters_temp,
  matrix_clusters
)
}
\arguments{
\item{mini_post}{numeric. A minimum to filter the significant clusters a posteriori. The default NULL is for no filtering with a a posteriori minimum.}

\item{maxi_post}{numeric. A maximum to filter the significant clusters a posteriori. The default NULL is for no filtering with a a posteriori maximum.}

\item{nb_sites}{numeric. The number of sites/individuals.}

\item{index_clusters_temp}{numeric vector. The indices of the detected clusters.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. A value of 1 indicate that the site (or the individual) is in the cluster, 0 otherwise.}
}
\value{
The detecting clusters with the a posteriori filtering.
}
\description{
This function allows the a posteriori filtering on the number of sites/individuals.
}
