% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Computation of the multivariate functional ranks}
\usage{
transform_data(data)
}
\arguments{
\item{data}{List. List of the data, each element of the list corresponds to a site (or an individual), each row corresponds to a variable and each column represents an observation time.}
}
\value{
List
}
\description{
This function computes the multivariate ranks of the data for each observation time
}
