% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_fonc.R
\name{uni_signs_matrix}
\alias{uni_signs_matrix}
\title{Computation of the matrix of signs}
\usage{
uni_signs_matrix(data)
}
\arguments{
\item{data}{numeric matrix. Matrix of the data, the rows correspond to the sites (or the individuals) and each column represents an observation time.}
}
\value{
numeric matrix.
}
\description{
This function returns the matrix of signs of the data.
}
