% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_uni.R
\name{wmw_uni}
\alias{wmw_uni}
\title{Index for the UNP scan procedure}
\usage{
wmw_uni(rank_data, matrix_clusters)
}
\arguments{
\item{rank_data}{matrix. Matrix of the ranks of the data for all permutations. Each column corresponds to a permutation and each row corresponds to a site or an individual.}

\item{matrix_clusters}{numeric matrix. Matrix in which each column represents a potential cluster. It is the result of the "clusters" function.}
}
\value{
numeric matrix.
}
\description{
This function returns the index we want to maximize on the set of potential clusters, for each potential cluster, and each permutation
}
